code fixed

+++ src/include/stir/scatter/ScatterSimulation.h
@@ -215,7 +215,7 @@ public:
     float detection_efficiency_integral(const float incoming_photon_energy, const float LT, const float HT, const float FWHM) const;
     float detection_efficiency_integral(const float incoming_photon_energy, const int en_window = 0) const;
     float detection_efficiency(const float incoming_photon_energy, const int en_window = 0) const;
-    std::vector<float> detection_spectrum(const float LLD, const float HLD, const float size, const float incoming_photon_energy) const;
+    std::vector<double> detection_spectrum(const float LLD, const float HLD, const float size, const float incoming_photon_energy) const;
     float detection_model_with_fitted_parameters(const float x, const float energy) const;
     float photoelectric(const float K, const float std_peak, const float x, const float energy) const;
     float compton_plateau(const float K, const float std_peak, const float x, const float energy, const float scaling_std_compton,const float shift_compton) const;
